Add files via upload
i have used linear regression on the IRI dataset The Iris dataset was used in R.A. Fisher's classic 1936 paper, The Use of Multiple Measurements in Taxonomic Problems, and can also be found on the UCI Machine Learning Repository. It includes three iris species with 50 samples each as well as some properties about each flower. The 3 species of iris areIris setosa, Iris virginica and Iris versicolor.